requirements.txt
Django==1.8.1
djangorestframework==3.3.3
psycopg2==2.6.1
wheel==0.24.0
